fix(backfill): preserve fractional share quantities, don't truncate to int
ClosedPositionRow.quantity was typed `int` and load_candidates cast
`int(row[5])`, truncating fractional positions to 0 or 1. This masked
realized P&L on every sub-1-share close — confirmed yesterday with
4 reconciliation_mismatch backfills on 5-12:

  XLC 1.9245 → qty=1, pnl=-$0.23 (real: -$0.44)
  XLF 1.989  → qty=1, pnl=-$0.23 (real: -$0.46)
  XLK 0.3927 → qty=0, pnl=-$0.00 (real: -$1.29)
  XLK 0.3127 → qty=0, pnl=-$0.00 (real: -$0.86)

The bot trades fractionals down to 1/1000000 (see CLAUDE.md), so this
silently hid >$2.50 of realized loss in a single day. Equity moved
-$3.01; daily_summaries showed $0 net.

Changes:
- quantity: int → float on ClosedPositionRow
- int(row[5]) → float(row[5]) in load_candidates
- log format qty=%d → qty=%.4f
- new test asserts a 0.3927-share position round-trips through
  load_candidates + backfill with full precision and the correct
  (exit - entry) * 0.3927 pnl

+++ trading_bot/self_improve/alpaca_backfill.py
@@ -55,7 +55,7 @@ class ClosedPositionRow:
     exchange: str
     currency: str
     strategy_id: str
-    quantity: int
+    quantity: float
     entry_price: float
     entry_time: datetime
     hold_type: str
@@ -138,7 +138,10 @@ def load_candidates(conn: sqlite3.Connection) -> list[ClosedPositionRow]:
                 exchange=row[2],
                 currency=row[3],
                 strategy_id=row[4],
-                quantity=int(row[5]),
+                # positions.quantity is INTEGER affinity but the live writer
+                # stores fractional shares (e.g. 0.3927). int() truncates
+                # those to 0, masking pnl on any sub-1-share position.
+                quantity=float(row[5]),
                 entry_price=float(row[6]),
                 entry_time=entry_dt,
                 hold_type=row[8],
@@ -357,7 +360,7 @@ async def backfill(
         gross_pnl = (exit_fill.filled_avg_price - c.entry_price) * c.quantity
         exit_reason = _infer_exit_reason(exit_fill.order_id, c)
         logger.info(
-            "Position %d %s qty=%d entry=%.4f exit=%.4f pnl=%+.2f reason=%s%s",
+            "Position %d %s qty=%.4f entry=%.4f exit=%.4f pnl=%+.2f reason=%s%s",
             c.position_id, c.ticker, c.quantity,
             c.entry_price, exit_fill.filled_avg_price, gross_pnl, exit_reason,
             " (dry-run)" if dry_run else "",
